Add a hash update timer to prevent excessive changes
Before this commit, any typed changes will cause the hash to change,
thus pushing a new URL to the user's browser history. Given that most
searches will be for full terms and not individual characters, this
commit addresses that issue by implementing a timer to update only
after a quarter-second of paused input to allow more meaningful hash
chunks.