Change site architecture to use Jekyll
This removes Node.js as a dependency for building the site, and allows
standalone updates to the sites.json file to render to the site without
doing any building of HTML files.

To account for these differences, the CI build has changed to utilize
a mixture of Ruby packages/scripts and a single call to a NPM package
(via `npx`) to perform similar validations and linting.

Written in collaboration with PyroSamurai <pyrosamurai@member.fsf.org>

Closes jdm-contrib/justdelete.me#194