Update instructions for MongoDB Atlas
The deletion process has changed and no longer requires one to send an
email to MongoDB's support team to delete your account. It can be done
from the site's settings page. Thus I've updated the difficulty and the
notes.

I've also changed the name from "Mongo Atlas" to "MongoDB Atlas" because
that's the name of the service according to their website.