Update sites.json
I considered this 'easy' because you don't have to email them and the button for deleting account obvious. This does state that it could take 48 hours to complete, but when I deleted my account it was immediately deleted and I was unable to log back into the account. The label 'medium' might make sense too because of the 48 hours potential waiting time.